+++ apps/block_scout_web/lib/block_scout_web/controllers/api/v2/transaction_controller.ex
@@ -304,8 +304,8 @@ defmodule BlockScoutWeb.API.V2.TransactionController do
   def optimism_batch(conn, %{"batch_number" => batch_number_string} = params) do
     {batch_number, ""} = Integer.parse(batch_number_string)
 
-    l2_block_number_from = OptimismTransactionBatch.edge_l2_block_number(batch_number, :min)
-    l2_block_number_to = OptimismTransactionBatch.edge_l2_block_number(batch_number, :max)
+    l2_block_number_from = OptimismTransactionBatch.edge_l2_block_number(batch_number, :min, @api_true)
+    l2_block_number_to = OptimismTransactionBatch.edge_l2_block_number(batch_number, :max, @api_true)
 
     handle_block_range_transactions(conn, params, l2_block_number_from, l2_block_number_to)
   end