+++ apps/block_scout_web/test/block_scout_web/controllers/api/v2/address_controller_test.exs
@@ -5496,8 +5496,8 @@ defmodule BlockScoutWeb.API.V2.AddressControllerTest do
     end
   end
 
-  describe "/addresses/{address_hash}/beacon/deposits" do
-    if Application.compile_env(:explorer, :chain_type) == :ethereum do
+  if @chain_type == :ethereum do
+    describe "/addresses/{address_hash}/beacon/deposits" do
       test "get empty list on non-existing address", %{conn: conn} do
         address = build(:address)
 
@@ -5530,16 +5530,12 @@ defmodule BlockScoutWeb.API.V2.AddressControllerTest do
         request = get(conn, "/api/v2/addresses/#{address.hash}/beacon/deposits")
         assert response = json_response(request, 200)
 
-        request_2nd_page = get(conn, "/api/v2/addresses/#{address.hash}/beacon/deposits", response["next_page_params"])
+        request_2nd_page =
+          get(conn, "/api/v2/addresses/#{address.hash}/beacon/deposits", response["next_page_params"])
+
         assert response_2nd_page = json_response(request_2nd_page, 200)
 
         check_paginated_response(response, response_2nd_page, deposits)
-      end
-    else
-      test "returns an error about chain type", %{conn: conn} do
-        request = get(conn, "/api/v2/addresses/address/beacon/deposits")
-        assert response = json_response(request, 404)
-        assert %{"message" => "Endpoint not available for current chain type"} = response
       end
     end
   end