+++ apps/block_scout_web/test/block_scout_web/controllers/api/v2/transaction_controller_test.exs
@@ -2086,8 +2086,8 @@ defmodule BlockScoutWeb.API.V2.TransactionControllerTest do
     end
   end
 
-  describe "transactions/{transaction_hash}/beacon/deposits" do
-    if Application.compile_env(:explorer, :chain_type) == :ethereum do
+  if @chain_type == :ethereum do
+    describe "transactions/{transaction_hash}/beacon/deposits" do
       test "get 404 on non-existing transaction", %{conn: conn} do
         transaction = build(:transaction)
 
@@ -2111,13 +2111,6 @@ defmodule BlockScoutWeb.API.V2.TransactionControllerTest do
         assert response_2nd_page = json_response(request_2nd_page, 200)
 
         check_paginated_response(response, response_2nd_page, deposits)
-      end
-    else
-      test "returns an error about chain type", %{conn: conn} do
-        transaction = insert(:transaction)
-        request = get(conn, "/api/v2/transactions/#{transaction.hash}/beacon/deposits")
-        assert response = json_response(request, 404)
-        assert %{"message" => "Endpoint not available for current chain type"} = response
       end
     end
   end
@@ -2930,6 +2923,8 @@ defmodule BlockScoutWeb.API.V2.TransactionControllerTest do
   end
 
   if @chain_type == :neon do
+    import Ecto.Query, only: [from: 2]
+
     describe "neon linked transactions service" do
       test "fetches data from the node and caches in the db", %{conn: conn} do
         transaction = insert(:transaction)