+++ apps/explorer/test/explorer/chain/token_transfer_test.exs
@@ -345,7 +345,8 @@ defmodule Explorer.Chain.TokenTransferTest do
               index: 0
             ),
           block: block,
-          address_hash: address.hash
+          address_hash: address.hash,
+          address: address
         )
 
       block_number = log.block_number
@@ -382,7 +383,7 @@ defmodule Explorer.Chain.TokenTransferTest do
         |> to_string()
         |> String.replace_prefix("0x000000000000000000000000", "0x")
 
-      token_contract_address = insert(:address, hash: log.address_hash)
+      token_contract_address = log.address
       to_address = insert(:address, hash: to_address_hash)
       from_address = insert(:address, hash: from_address_hash)