+++ apps/indexer/test/indexer/temporary/uncataloged_token_transfers_test.exs
@@ -46,6 +46,7 @@ defmodule Indexer.Temporary.UncatalogedTokenTransfersTest do
         insert(:token_transfer_log,
           transaction: transaction,
           address_hash: address.hash,
+          address: address,
           block: block
         )